pyproject.toml
[tool.poetry]
name = "Password Generator"
version = "1.9.0"
description = "Generate secure passwords"
documentation = "https://github.com/fedecalendino/alfred-pwd-gen/blob/main/README.md"
homepage = "https://github.com/fedecalendino/alfred-pwd-gen"
license = "MIT"

authors = [
    "Fede Calendino <fede@calendino.com>"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
alfred-pyflow = "0.4.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
coverage = "^7.2.5"
ddt = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
